(refactoring) Split TEvChangeExchange into two parts: common & DS KIKIMR-20673
