+++ ydb/core/tx/datashard/change_sender_cdc_stream.cpp
@@ -83,14 +83,14 @@ class TCdcChangeSenderPartition: public TActorBootstrapped<TCdcChangeSenderParti
 
     STATEFN(StateWaitingRecords) {
         switch (ev->GetTypeRewrite()) {
-            hFunc(TEvChangeExchange::TEvRecords, Handle);
+            hFunc(NChangeExchange::TEvChangeExchange::TEvRecords, Handle);
             sFunc(TEvPartitionWriter::TEvWriteResponse, Lost);
         default:
             return StateBase(ev);
         }
     }
 
-    void Handle(TEvChangeExchange::TEvRecords::TPtr& ev) {
+    void Handle(NChangeExchange::TEvChangeExchange::TEvRecords::TPtr& ev) {
         LOG_D("Handle " << ev->Get()->ToString());
         NKikimrClient::TPersQueueRequest request;
 
@@ -703,17 +703,17 @@ class TCdcChangeSenderMain
         return new TCdcChangeSenderPartition(SelfId(), DataShard, partitionId, shardId, Stream);
     }
 
-    void Handle(TEvChangeExchange::TEvEnqueueRecords::TPtr& ev) {
+    void Handle(NChangeExchange::TEvChangeExchange::TEvEnqueueRecords::TPtr& ev) {
         LOG_D("Handle " << ev->Get()->ToString());
         EnqueueRecords(std::move(ev->Get()->Records));
     }
 
-    void Handle(TEvChangeExchange::TEvRecords::TPtr& ev) {
+    void Handle(NChangeExchange::TEvChangeExchange::TEvRecords::TPtr& ev) {
         LOG_D("Handle " << ev->Get()->ToString());
         ProcessRecords(std::move(ev->Get()->Records));
     }
 
-    void Handle(TEvChangeExchange::TEvForgetRecords::TPtr& ev) {
+    void Handle(NChangeExchange::TEvChangeExchange::TEvForgetRecords::TPtr& ev) {
         LOG_D("Handle " << ev->Get()->ToString());
         ForgetRecords(std::move(ev->Get()->Records));
     }
@@ -736,7 +736,7 @@ class TCdcChangeSenderMain
         PassAway();
     }
 
-    void AutoRemove(TEvChangeExchange::TEvEnqueueRecords::TPtr& ev) {
+    void AutoRemove(NChangeExchange::TEvChangeExchange::TEvEnqueueRecords::TPtr& ev) {
         LOG_D("Handle " << ev->Get()->ToString());
         RemoveRecords(std::move(ev->Get()->Records));
     }
@@ -768,9 +768,9 @@ public:
 
     STFUNC(StateBase) {
         switch (ev->GetTypeRewrite()) {
-            hFunc(TEvChangeExchange::TEvEnqueueRecords, Handle);
-            hFunc(TEvChangeExchange::TEvRecords, Handle);
-            hFunc(TEvChangeExchange::TEvForgetRecords, Handle);
+            hFunc(NChangeExchange::TEvChangeExchange::TEvEnqueueRecords, Handle);
+            hFunc(NChangeExchange::TEvChangeExchange::TEvRecords, Handle);
+            hFunc(NChangeExchange::TEvChangeExchange::TEvForgetRecords, Handle);
             hFunc(TEvChangeExchange::TEvRemoveSender, Handle);
             hFunc(TEvChangeExchangePrivate::TEvReady, Handle);
             hFunc(TEvChangeExchangePrivate::TEvGone, Handle);
@@ -781,7 +781,7 @@ public:
 
     STFUNC(StatePendingRemove) {
         switch (ev->GetTypeRewrite()) {
-            hFunc(TEvChangeExchange::TEvEnqueueRecords, AutoRemove);
+            hFunc(NChangeExchange::TEvChangeExchange::TEvEnqueueRecords, AutoRemove);
             hFunc(TEvChangeExchange::TEvRemoveSender, Handle);
             HFunc(NMon::TEvRemoteHttpInfo, Handle);
             sFunc(TEvents::TEvPoison, PassAway);
